Added Patient Name screen between main and adding new device
