+++ app/src/main/java/com/mbientlab/metawear/app/MainActivity.java
@@ -10,7 +10,6 @@ import android.bluetooth.BluetoothDevice;
 import android.content.Intent;
 
 public class MainActivity extends AppCompatActivity {
-    public static final int REQUEST_START_BLE_SCAN= 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -20,19 +19,10 @@ public class MainActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
 
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(view -> startActivityForResult(new Intent(MainActivity.this, ScannerActivity.class), REQUEST_START_BLE_SCAN));
-    }
+        fab.setOnClickListener(view -> startActivity(new Intent(MainActivity.this, PatientName.class)));
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        switch (requestCode) {
-            case REQUEST_START_BLE_SCAN:
-                BluetoothDevice selectedDevice= data.getParcelableExtra(ScannerActivity.EXTRA_DEVICE);
-                if (selectedDevice != null) {
-                    ((MainActivityFragment) getSupportFragmentManager().findFragmentById(R.id.main_activity_content)).addNewDevice(selectedDevice);
-                }
-                break;
-        }
+        //FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
+        //fab.setOnClickListener(view -> startActivityForResult(new Intent(MainActivity.this, ScannerActivity.class), REQUEST_START_BLE_SCAN));
     }
 
 }
