+++ app/src/main/res/menu/activity_navigation_drawer.xml
@@ -26,7 +26,8 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <group android:checkableBehavior="single">
         <item android:id="@+id/nav_home" android:title="@string/navigation_fragment_home" android:checked="true" />
-        <item android:id="@+id/nav_accelerometer" android:title="@string/navigation_fragment_accelerometer" />
+        <item android:id="@+id/nav_sensor_fusion" android:title="@string/navigation_fragment_sensor_fusion" />
+        <!--<item android:id="@+id/nav_accelerometer" android:title="@string/navigation_fragment_accelerometer" />
         <item android:id="@+id/nav_barometer" android:title="@string/navigation_fragment_barometer" />
         <item android:id="@+id/nav_color_detector" android:title="@string/navigation_fragment_color_detector" />
         <item android:id="@+id/nav_gpio" android:title="@string/navigation_fragment_gpio" />
@@ -41,6 +42,6 @@
         <item android:id="@+id/nav_proximity" android:title="@string/navigation_fragment_proximity" />
         <item android:id="@+id/nav_sensor_fusion" android:title="@string/navigation_fragment_sensor_fusion" />
         <item android:id="@+id/nav_settings" android:title="@string/navigation_fragment_settings" />
-        <item android:id="@+id/nav_temperature" android:title="@string/navigation_fragment_temperature" />
+        <item android:id="@+id/nav_temperature" android:title="@string/navigation_fragment_temperature" />-->
     </group>
 </menu>
