+++ app/src/main/res/values/strings.xml
@@ -304,4 +304,7 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="title_activity_patient_name">PatientName</string>
+    <string name="prompt_name">Enter Patient Name</string>
+    <string name="add_patient">Add</string>
 </resources>
